Fixing ReadBackedPileup to represent mapping qualities as ints, not (signed) bytes. Having them as bytes caused problems for downstream programmers who had data with high MQs.

+++ protected/java/src/org/broadinstitute/sting/gatk/walkers/diagnostics/missing/QualifyMissingIntervals.java
@@ -156,7 +156,7 @@ public final class QualifyMissingIntervals extends LocusWalker<Metrics, Metrics>
             baseQual += qual;
         }
         double mapQual = 0.0;
-        for (byte qual : pileup.getMappingQuals()) {
+        for (int qual : pileup.getMappingQuals()) {
             mapQual += qual;
         }
